+++ mac/mac.h
@@ -32,10 +32,10 @@ typedef struct
 mac_header_t;
 
 
-int mac_header_write (mac_header_t* mac_header, uint8_t tx_buf[], uint16_t tx_buf_len);
+int16_t mac_header_write (mac_header_t* mac_header, uint8_t tx_buf[], uint16_t tx_buf_len);
 
 
-int mac_header_read (mac_header_t* mac_header, uint8_t rx_buf[], uint16_t rx_buf_len);
+int16_t mac_header_read (mac_header_t* mac_header, uint8_t rx_buf[], uint16_t rx_buf_len);
 
 
 #endif
