+++ port/deca_sleep.c
@@ -10,7 +10,7 @@
 
 
 // Sleep for time_ms milliseconds
-void deca_sleep (unsigned int sleep_time_ms)
+void deca_sleep (uint16_t sleep_time_ms)
 {
     k_msleep(sleep_time_ms);
 
@@ -19,7 +19,7 @@ void deca_sleep (unsigned int sleep_time_ms)
 
 
 // Sleep for time_us microseconds
-void deca_usleep (unsigned long sleep_time_us)
+void deca_usleep (uint32_t sleep_time_us)
 {
     k_usleep(sleep_time_us);
